+++ include/pagmo/algorithms/pso_gen.hpp
@@ -33,7 +33,10 @@ see https://www.gnu.org/licenses/. */
 #include <tuple>
 #include <vector>
 
+#include <boost/optional.hpp>
+
 #include <pagmo/algorithm.hpp>
+#include <pagmo/bfe.hpp>
 #include <pagmo/detail/visibility.hpp>
 #include <pagmo/population.hpp>
 #include <pagmo/rng.hpp>
@@ -150,6 +153,8 @@ public:
     {
         return m_verbosity;
     }
+    // Sets the bfe
+    void set_bfe(const bfe &b);
     // Sets the seed
     void set_seed(unsigned);
     /// Gets the seed
@@ -220,6 +225,7 @@ private:
     unsigned m_seed;
     unsigned m_verbosity;
     mutable log_type m_log;
+    boost::optional<bfe> m_bfe;
 };
 
 } // namespace pagmo